# [BesTLA] AVX2: Use loaded registers of B.